Pencil sketch with python
The “Pencil Sketch with Python” project is a fascinating exploration of computer vision techniques using Python. In this project, an image is transformed into a pencil sketch using a few simple steps:

Image Reading: The original image is read into the program using a library such as OpenCV.
Grayscale Conversion: The image is then converted to grayscale, as pencil sketches are monochrome.
Inversion: The grayscale image is inverted, meaning that the colors are flipped. Light regions become dark and vice versa.
Blurring: The inverted image is blurred using a Gaussian function. This helps to create the shading effect of the sketch.
Final Sketch Creation: The blurred, inverted image is blended with the grayscale image to create the final pencil sketch.